+++ llvm-hs/src/LLVM/Internal/FFI/BuilderC.cpp
@@ -61,7 +61,10 @@ LLVMValueRef LLVM_Hs_Build ## Op(																	\
 	LLVMValueRef o1,																											\
 	const char *s																													\
 ) {																																			\
-	return wrap(unwrap(b)->Create ## Op(unwrap(o0), unwrap(o1), s, nuw, nsw)); \
+    BinaryOperator *BO = unwrap(b)->Insert(BinaryOperator::Create(Instruction::Op, unwrap(o0), unwrap(o1)), s); \
+    if (nuw) BO->setHasNoUnsignedWrap(); \
+    if (nsw) BO->setHasNoSignedWrap(); \
+    return wrap(BO);                   \
 }
 LLVM_HS_FOR_EACH_OVERFLOWING_BINARY_OPERATOR(ENUM_CASE)
 #undef ENUM_CASE
@@ -74,13 +77,74 @@ LLVMValueRef LLVM_Hs_Build ## Op(																	\
 	LLVMValueRef o1,																											\
 	const char *s																													\
 ) {																																			\
-	return wrap(unwrap(b)->Create ## Op(unwrap(o0), unwrap(o1), s, exact)); \
+    if (!exact) { \
+        return wrap(unwrap(b)->Insert(BinaryOperator::Create##Op(unwrap(o0), unwrap(o1)), s)); \
+    } \
+     return wrap(unwrap(b)->Insert(BinaryOperator::CreateExact##Op(unwrap(o0), unwrap(o1)), s)); \
 }
 LLVM_HS_FOR_EACH_POSSIBLY_EXACT_BINARY_OPERATOR(ENUM_CASE)
 #undef ENUM_CASE
 
+#define LLVM_HS_FOR_EACH_CAST_INSTR(macro) \
+    macro(AddrSpaceCast) \
+    macro(BitCast) \
+    macro(FPExt) \
+    macro(FPToSI) \
+    macro(FPToUI) \
+    macro(FPTrunc) \
+    macro(IntToPtr) \
+    macro(PtrToInt) \
+    macro(SExt) \
+    macro(SIToFP) \
+    macro(Trunc) \
+    macro(UIToFP) \
+    macro(ZExt)
+
+#define ENUM_CASE(Op) \
+LLVMValueRef LLVM_Hs_Build##Op(LLVMBuilderRef b, LLVMValueRef val, LLVMTypeRef destTy, const char *name) {\
+    return wrap(unwrap(b)->Insert(CastInst::Create(Instruction::Op, unwrap(val), unwrap(destTy)), name));\
+}
+LLVM_HS_FOR_EACH_CAST_INSTR(ENUM_CASE)
+#undef ENUM_CASE
+
+#define LLVM_HS_FOR_EACH_BINOP(macro) \
+    macro(And) \
+    macro(Or) \
+    macro(SRem) \
+    macro(URem) \
+    macro(Xor)
+
+#define ENUM_CASE(Op) \
+LLVMValueRef LLVM_Hs_Build##Op(LLVMBuilderRef b, LLVMValueRef lhs, LLVMValueRef rhs, const char *name) {\
+    return wrap(unwrap(b)->Insert(BinaryOperator::Create##Op(unwrap(lhs), unwrap(rhs)), name)); \
+}
+LLVM_HS_FOR_EACH_BINOP(ENUM_CASE)
+#undef ENUM_CASE
+
+#define LLVM_HS_FOR_EACH_FP_BINOP(macro) \
+    macro(FAdd) \
+    macro(FDiv) \
+    macro(FMul) \
+    macro(FRem) \
+    macro(FSub)
+
+#define ENUM_CASE(Op) \
+LLVMValueRef LLVM_Hs_Build##Op(LLVMBuilderRef b, LLVMValueRef lhs, LLVMValueRef rhs, const char *name) { \
+    BinaryOperator* bo = BinaryOperator::Create##Op(unwrap(lhs), unwrap(rhs)); \
+    bo->setFastMathFlags(unwrap(b)->getFastMathFlags()); \
+    return wrap(unwrap(b)->Insert(bo, name)); \
+}
+LLVM_HS_FOR_EACH_FP_BINOP(ENUM_CASE)
+#undef ENUM_CASE
+
 void LLVM_Hs_SetFastMathFlags(LLVMBuilderRef b, LLVMFastMathFlags f) {
 	unwrap(b)->setFastMathFlags(unwrap(f));
+}
+
+LLVMValueRef LLVM_Hs_BuildICmp(LLVMBuilderRef b, LLVMIntPredicate op, LLVMValueRef lhs, LLVMValueRef rhs, const char* name) {
+
+    return wrap(unwrap(b)->Insert(new ICmpInst(static_cast<ICmpInst::Predicate>(op),
+                                               unwrap(lhs), unwrap(rhs)), name));
 }
 
 LLVMValueRef LLVM_Hs_BuildLoad(
@@ -160,27 +224,6 @@ LLVMValueRef LLVM_Hs_BuildAtomicRMW(
 	return wrap(a);
 }
 
-LLVMValueRef LLVM_Hs_BuildExtractValue(
-	LLVMBuilderRef b,
-	LLVMValueRef a,
-	unsigned *idxs,
-	unsigned n,
-	const char *name
-) {
-	return wrap(unwrap(b)->CreateExtractValue(unwrap(a), ArrayRef<unsigned>(idxs, n), name));
-}
-
-LLVMValueRef LLVM_Hs_BuildInsertValue(
-	LLVMBuilderRef b,
-	LLVMValueRef a,
-	LLVMValueRef v,
-	unsigned *idxs,
-	unsigned n,
-	const char *name
-) {
-	return wrap(unwrap(b)->CreateInsertValue(unwrap(a), unwrap(v), ArrayRef<unsigned>(idxs, n), name));
-}
-
 LLVMValueRef LLVM_Hs_BuildCleanupPad(LLVMBuilderRef b, LLVMValueRef parentPad,
                                      LLVMValueRef *args, unsigned numArgs,
                                      const char *name) {
@@ -221,5 +264,68 @@ LLVMValueRef LLVM_Hs_BuildCatchSwitch(LLVMBuilderRef b, LLVMValueRef parentPad,
                                       unsigned numHandlers) {
     return wrap(unwrap(b)->CreateCatchSwitch(unwrap(parentPad),
                                              unwrap(unwindDest), numHandlers));
+}
+
+LLVMValueRef LLVM_Hs_BuildGEP(LLVMBuilderRef B, LLVMValueRef Pointer,
+                             LLVMValueRef *Indices, unsigned NumIndices,
+                             const char *Name) {
+  ArrayRef<Value *> IdxList(unwrap(Indices), NumIndices);
+  return wrap(unwrap(B)->Insert(GetElementPtrInst::Create(nullptr, unwrap(Pointer), IdxList), Name));
+}
+
+LLVMValueRef LLVM_Hs_BuildInBoundsGEP(LLVMBuilderRef B, LLVMValueRef Pointer,
+                                      LLVMValueRef *Indices, unsigned NumIndices,
+                                      const char *Name) {
+  ArrayRef<Value *> IdxList(unwrap(Indices), NumIndices);
+  return wrap(unwrap(B)->Insert(GetElementPtrInst::CreateInBounds(nullptr, unwrap(Pointer), IdxList), Name));
+}
+
+LLVMValueRef LLVM_Hs_BuildSelect(LLVMBuilderRef B, LLVMValueRef If,
+                                 LLVMValueRef Then, LLVMValueRef Else,
+                                 const char *Name) {
+    return wrap(unwrap(B)->Insert(SelectInst::Create(unwrap(If), unwrap(Then), unwrap(Else))));
+}
+
+LLVMValueRef LLVM_Hs_BuildExtractValue(
+	LLVMBuilderRef b,
+	LLVMValueRef a,
+	unsigned *idxs,
+	unsigned n,
+	const char *name
+) {
+	return wrap(unwrap(b)->Insert(ExtractValueInst::Create(unwrap(a), ArrayRef<unsigned>(idxs, n)), name));
+}
+
+LLVMValueRef LLVM_Hs_BuildInsertValue(
+	LLVMBuilderRef b,
+	LLVMValueRef a,
+	LLVMValueRef v,
+	unsigned *idxs,
+	unsigned n,
+	const char *name
+) {
+	return wrap(unwrap(b)->Insert(InsertValueInst::Create(unwrap(a), unwrap(v), ArrayRef<unsigned>(idxs, n)), name));
+}
+
+LLVMValueRef LLVM_Hs_BuildExtractElement(LLVMBuilderRef B, LLVMValueRef VecVal,
+                                         LLVMValueRef Index, const char *Name) {
+    return wrap(unwrap(B)->Insert(ExtractElementInst::Create(unwrap(VecVal), unwrap(Index)),
+                                  Name));
+}
+
+LLVMValueRef LLVM_Hs_BuildInsertElement(LLVMBuilderRef B, LLVMValueRef VecVal,
+                                        LLVMValueRef EltVal, LLVMValueRef Index,
+                                        const char *Name) {
+    return wrap(unwrap(B)->Insert(InsertElementInst::Create(unwrap(VecVal), unwrap(EltVal),
+                                                            unwrap(Index)),
+                                  Name));
+}
+
+LLVMValueRef LLVM_Hs_BuildShuffleVector(LLVMBuilderRef B, LLVMValueRef V1,
+                                        LLVMValueRef V2, LLVMValueRef Mask,
+                                        const char *Name) {
+    return wrap(unwrap(B)->Insert(new ShuffleVectorInst(unwrap(V1), unwrap(V2),
+                                                        unwrap(Mask)),
+                                  Name));
 }
 }