+++ llvm-hs/src/LLVM/Internal/ObjectFile.hs
@@ -1,5 +1,6 @@
 module LLVM.Internal.ObjectFile where
 
+import Control.Exception (bracket)
 import Control.Monad.IO.Class
 import Control.Monad.AnyCont
 import Foreign.Ptr
@@ -7,6 +8,7 @@ import Foreign.Ptr
 import LLVM.Prelude
 import LLVM.Internal.Coding
 import LLVM.Internal.MemoryBuffer
+import qualified LLVM.Internal.FFI.LLVMCTypes as FFI
 import qualified LLVM.Internal.FFI.ObjectFile as FFI
 
 newtype ObjectFile = ObjectFile (Ptr FFI.ObjectFile)
@@ -22,7 +24,13 @@ disposeObjectFile (ObjectFile obj) = FFI.disposeObjectFile obj
 -- `.o` file. This does *not* compile source files.
 createObjectFile :: FilePath -> IO ObjectFile
 createObjectFile path = flip runAnyContT return $ do
-  buf <- encodeM (File path)
+  -- The ownership of the object file is transfered to the object
+  -- file, so we need to make sure that we do not free it here.
+  FFI.OwnerTransfered buf <- encodeM (File path)
   obj <- liftIO $ FFI.createObjectFile buf
   when (obj == nullPtr) $ error "LLVMCreateObjectFile returned a null pointer."
   return (ObjectFile obj)
+
+-- | @bracket@-style wrapper for `createObjectFile` and `disposeObjectFile`.
+withObjectFile :: FilePath -> (ObjectFile -> IO a) -> IO a
+withObjectFile f = bracket (createObjectFile f) disposeObjectFile